Manual

Quickstart

While StateSpaceModels.jl offers a rich array of functionalities, diverse models, and flexible interfaces, most users primarily seek to achieve core tasks: fitting a model and generating forecasts. The following example provides a concise introduction to performing these essential operations effectively.

using StateSpaceLearning

y = randn(100)

# Instantiate Model
model = StructuralModel(y)

# Fit Model
fit!(model)

# Point Forecast
prediction = StateSpaceLearning.forecast(model, 12) #Gets a 12 steps ahead prediction

# Scenarios Path Simulation
simulation = StateSpaceLearning.simulate(model, 12, 1000) #Gets 1000 scenarios path of 12 steps ahead predictions

Models

The package currently supports the implementation of the StructuralModel. If you have suggestions for additional models to include, we encourage you to contribute by opening an issue or submitting a pull request.

Main.StateSpaceLearning.StructuralModel — Type

Instantiates a Structural State Space Learning model.

StructuralModel(
    y::Union{Vector,Matrix};
    level::Bool=true,
    stochastic_level::Bool=true,
    trend::Bool=true,
    stochastic_trend::Bool=true,
    seasonal::Bool=true,
    stochastic_seasonal::Bool=true,
    freq_seasonal::Union{Int, Vector{Int}}=12,
    outlier::Bool=true,
    ζ_ω_threshold::Int=12,
    stochastic_start::Int=1,
    Exogenous_X::Matrix=if typeof(y) <: Vector
        zeros(length(y), 0)
    else
        zeros(size(y, 1), 0)
    end,
)

A Structural State Space Learning model that can have level, stochasticlevel, trend, stochastictrend, seasonal, stochastic_seasonal, outlier and Exogenous components. Each component should be specified by Booleans.

These models take the general form:

\[\begin{gather*} \begin{aligned} y_1 &= \mu_1 + \gamma_1 + X_1\beta + \sum_{\tau=1}^1D_{\tau,1} \,o_\tau + \varepsilon_1\\ y_2 &= \mu_1 + \xi_2 + \nu_1 + \gamma_2 + X_2\beta + \sum_{\tau=1}^2D_{\tau,2} \,o_\tau + \varepsilon_2 \\ y_t &= \mu_1 + \sum_{\tau=2}^{t}\xi_\tau + (t-1)\nu_1 + \sum_{\tau=2}^{t-1}(t-\tau)\zeta_\tau + \gamma_{m_t} + X_t\beta + \sum_{\tau=1}^tD_{\tau,t} \,o_\tau + \varepsilon_t, \quad \forall t = \{3, \ldots, s\} \label{cor_t2}\\ y_t &= \mu_1 + \sum_{\tau=2}^{t}\xi_\tau + (t-1)\nu_1 + \sum_{\tau=2}^{t-1}(t-\tau)\zeta_\tau + \gamma_{m_{t}} + \sum_{\tau \in M_{t}}(\omega_{{\tau}} - \omega_{\tau-1})+ X_t\beta + \sum_{\tau=1}^tD_{\tau,t} \,o_\tau + \varepsilon_t, \quad \forall t=s+1,\ldots,T \\ % \zeta_t, \xi_t, \omega_t =0, \; \forall \; t > T \end{aligned} \end{gather*}\]

The notation is as follows: $y_t$ represents the observation vector at time $t$, $\mu_1$ denotes the initial level component (intercept), and $\xi_t$ refers to the stochastic level component. Similarly, $\nu_1$ corresponds to the deterministic slope component, and $\zeta_t$ represents the stochastic slope component. The seasonal effects are described by $\gamma_{m_t}$ for the deterministic seasonal component and $\omega_{\tau}$ for the stochastic seasonal component.

The exogenous component is represented by $X$, with $\beta$ as the associated coefficients. Outlier effects are captured by the dummy outlier matrix $D$ and its corresponding coefficients $o$. Finally, $\varepsilon_t$ denotes the irregular term.

References

Ramos, André, & Valladão, Davi, & Street, Alexandre. Time Series Analysis by State Space Learning

Example

y = rand(100)
model = StructuralModel(y)

Fitting

The package currently only allows for the estimation procedure (based on the elastic net) presented in the paper "Time Series Analysis by State Space Learning". We allow some parameters configurations as detailed below. If you have suggestions for additional estimation procedures to include, we encourage you to contribute by opening an issue or submitting a pull request.

Main.StateSpaceLearning.fit! — Function

Fits the StateSpaceLearning model using specified parameters and estimation procedures. After fitting, the model output can be accessed in model.output

fit!(model::StateSpaceLearningModel, α::AbstractFloat = 0.1, information_criteria::String = "aic", ϵ::AbstractFloat = 0.05, penalize_exogenous::Bool = true, penalize_initial_states::Bool = true, )

Arguments

model::StateSpaceLearningModel: Model to be fitted.

α::AbstractFloat: Elastic net mixing parameter (default: 0.1).

information_criteria::String: Method for hyperparameter selection (default: "aic").

ϵ::AbstractFloat: Non negative value to handle 0 coefs on the first lasso step (default: 0.05).

penalize_exogenous::Bool: If true, penalize exogenous variables (default: true).

penalize_initial_states::Bool: If true, penalize initial states (default: true).

Example

y = rand(100)
model = StructuralModel(y)
fit!(model)
output = model.output

Forecasting and Simulating

The package has functions to make point forecasts multiple steps ahead and to simulate scenarios based on those forecasts. These functions are implemented both for the univariate and to the multivariate cases.

Main.StateSpaceLearning.forecast — Function

Returns the forecast for a given number of steps ahead using the provided StateSpaceLearning output and exogenous forecast data.

forecast(model::StateSpaceLearningModel, steps_ahead::Int; Exogenous_Forecast::Union{Matrix{Fl}, Missing}=missing)::Vector{AbstractFloat} where Fl

Arguments

model::StateSpaceLearningModel: Model obtained from fitting.

steps_ahead::Int: Number of steps ahead for forecasting.

Exogenous_Forecast::Matrix{Fl}: Exogenous variables forecast (default: zeros(steps_ahead, 0))

Returns

Union{Matrix{AbstractFloat}, Vector{AbstractFloat}}: Matrix or vector of matrices containing forecasted values.

Example

y = rand(100)
model = StructuralModel(y)
fit!(model)
steps_ahead = 12
point_prediction = forecast(model, steps_ahead)

Main.StateSpaceLearning.simulate — Function

Generate simulations for a given number of steps ahead using the provided StateSpaceLearning output and exogenous forecast data.

simulate(model::StateSpaceLearningModel, steps_ahead::Int, N_scenarios::Int; Exogenous_Forecast::Matrix{Fl}=zeros(steps_ahead, 0))::Matrix{AbstractFloat} where Fl

Arguments

model::StateSpaceLearningModel: Model obtained from fitting.

steps_ahead::Int: Number of steps ahead for simulation.

N_scenarios::Int: Number of scenarios to simulate (default: 1000).

Exogenous_Forecast::Matrix{Fl}: Exogenous variables forecast (default: zeros(steps_ahead, 0))

Returns

Union{Vector{Matrix{AbstractFloat}}, Matrix{AbstractFloat}}: Matrix or vector of matrices containing simulated values.

Example (Univariate Case)

y = rand(100)
model = StructuralModel(y)
fit!(model)
steps_ahead = 12
N_scenarios = 1000
simulation  = simulate(model, steps_ahead, N_scenarios)

Example (Multivariate Case)

y = rand(100, 3)
model = StructuralModel(y)
fit!(model)
steps_ahead = 12
N_scenarios = 1000
simulation  = simulate(model, steps_ahead, N_scenarios)

Datasets

The package includes several datasets designed to demonstrate its functionalities and showcase the models. These datasets are stored as CSV files, and their file paths can be accessed either by their names as shown below. In the examples, we utilize DataFrames.jl and CSV.jl to illustrate how to work with these datasets.

Main.StateSpaceLearning.AIR_PASSENGERS — Constant

AIR_PASSENGERS

The absolute path for the AIR_PASSENGERS dataset stored inside StateSpaceLearning.jl. This dataset provides monthly totals of a US airline passengers from 1949 to 1960.

References

https://www.stata-press.com/data/r12/ts.html

Main.StateSpaceLearning.ARTIFICIAL_SOLARS — Constant

ARTIFICIAL_SOLARS

The absolute path for the ARTIFICIAL_SOLARS dataset stored inside StateSpaceLearning.jl. This dataset provides an hourly Multivariate Time Series for 3 artificial solar power plants.

Main.StateSpaceLearning.HOURLY_M4_EXAMPLE — Constant

HOURLY_M4_EXAMPLE

The absolute path for the HOURLY_M4_EXAMPLE dataset stored inside StateSpaceLearning.jl. This dataset provides an hourly Time Series from the M4 competition dataset.

References

https://github.com/Mcompetitions/M4-methods